+++ shiva/day1/answer.c
