+++ shiva/day2/answer.c
