+++ shiva/day3/answer1.c
