+++ shiva/day3/answer2.c
