+++ shiva/day4/answer.c
