+++ shiva/day6/answer.c
