+++ shiva/day7/answer.c
