+++ shiva/day7/question.txt
