--- shiva/day8/question.txt
@@ -1,3 +1,0 @@
-This challenge will help you learn the concept of recursion.
-There is a series, s , where the next term is the sum of pervious three terms. Given the first three terms of the series, a, b, and c respectively, you have to output the nth term of the series using recursion.
-question taken from hackerrank
